Show Data Category

+++ resources/views/admin/category/index.blade.php
@@ -43,7 +43,7 @@
                                 <td>{{$rs->image}}</td>
                                 <td>{{$rs->status}}</td>
                                 <td><a href="/admin/category/edit/{{$rs->id}}" class="btn btn-info">Edit</a></td>
-                                <td><a href="/admin/category/delete/{{$rs->id}}"class="btn btn-danger">Delete</a></td>
+                                <td><a href="/admin/category/delete/{{$rs->id}}"onclick="return confirm('Deleting!! Are you sure?')"class="btn btn-danger">Delete</a></td>
                                 <td><a href="/admin/category/show/{{$rs->id}}"class="btn btn-success">Show</a></td>
                             </tr>
                             @endforeach
